// src/modules/container/domain/container.type.ts
export interface ContainerProps {
  id?: bigint | null;
  idTos?: string | null;
  idCont?: string | null;
  idContTos?: string | null;
  vesselKey: string;
  vesselImvoy?: string | null;
  vesselExvoy?: string | null;
  eta?: Date | null;
  etb?: Date | null;
  etd?: Date | null;
  bargeInKey?: string | null;
  bargeOutKey?: string | null;
  deliveryOrder?: string | null;
  blNo?: string | null;
  bookingNo?: string | null;
  houseBillNo?: string | null;
  containerNo: string;
  classCode: string;
  operationCode: string;
  fe: string;
  containerStatusCode: string;
  cargoTypeCode: string;
  commodity?: string | null;
  localSizeType: string;
  isoSizeType: string;
  isLocalForeign: string;
  jobModeCodeIn?: string | null;
  methodCodeIn?: string | null;
  dateIn?: Date | null;
  jobModeCodeOut?: string | null;
  methodCodeOut?: string | null;
  dateOut?: Date | null;
  requireSurveyDate?: Date | null;
  repairMoveDate?: Date | null;
  eirInNo?: string | null;
  eirOutNo?: string | null;
  stuffNo?: string | null;
  unstuffNo?: string | null;
  serviceNo?: string | null;
  draftNo?: string | null;
  invoiceNo?: string | null;
  zoneCode?: string | null;
  yardCode?: string | null;
  lineCode?: string | null;
  block?: string | null;
  bay?: string | null;
  row?: string | null;
  tier?: string | null;
  area?: string | null;
  vgm: boolean;
  mcWeight?: Prisma.Decimal | null;
  tareWeight?: Prisma.Decimal | null;
  maxGrossWeight?: Prisma.Decimal | null;
  sealNo?: string | null;
  sealNo1?: string | null;
  sealNo2?: string | null;
  pol?: string | null;
  pod?: string | null;
  fpod?: string | null;
  transitCode?: string | null;
  transitPort?: string | null;
  temperature?: string | null;
  vent?: Prisma.Decimal | null;
  ventUnit?: string | null;
  cusHold: boolean;
  terHold: boolean;
  terHoldReason?: string | null;
  isSpecialWarning: boolean;
  specialWarning?: string | null;
  conditionCode?: string | null;
  classifyCode?: string | null;
  conditionMachineCode?: string | null;
  isTruckBarge?: string | null;
  truckNo?: string | null;
  romoocNo?: string | null;
  isBundled?: number;
  containerNoMaster?: string | null;
  noteCont?: string | null;
  noteDamage?: string | null;
  noteSpecialHandling?: string | null;
  note?: string | null;
  createdBy: string;
  updatedBy?: string | null;
}

export interface CreateContainerProps {
  idTos?: string | null;
  idCont?: string | null;
  idContTos?: string | null;
  vesselKey: string;
  vesselImvoy?: string | null;
  vesselExvoy?: string | null;
  eta?: Date | null;
  etb?: Date | null;
  etd?: Date | null;
  bargeInKey?: string | null;
  bargeOutKey?: string | null;
  deliveryOrder?: string | null;
  blNo?: string | null;
  bookingNo?: string | null;
  housebillNo?: string | null;
  containerNo: string;
  classCode: string;
  operationCode: string;
  fe: string;
  containerStatusCode: string;
  cargoTypeCode: string;
  commodity?: string | null;
  localSizeType: string;
  isoSizeType: string;
  isLocalForeign: string;
  jobmodeCodeIn?: string | null;
  methodCodeIn?: string | null;
  dateIn?: Date | null;
  jobmodeCodeOut?: string | null;
  methodCodeOut?: string | null;
  dateOut?: Date | null;
  requireSurveyDate?: Date | null;
  repairMoveDate?: Date | null;
  eirInNo?: string | null;
  eirOutNo?: string | null;
  stuffNo?: string | null;
  unstuffNo?: string | null;
  serviceNo?: string | null;
  draftNo?: string | null;
  invoiceNo?: string | null;
  zoneCode?: string | null;
  yardCode?: string | null;
  lineCode?: string | null;
  block?: string | null;
  bay?: string | null;
  row?: string | null;
  tier?: string | null;
  area?: string | null;
  vgm: boolean;
  mcWeight?: Prisma.Decimal | null;
  tareWeight?: Prisma.Decimal | null;
  maxGrossWeight?: Prisma.Decimal | null;
  sealNo?: string | null;
  sealNo1?: string | null;
  sealNo2?: string | null;
  pol?: string | null;
  pod?: string | null;
  fpod?: string | null;
  transitCode?: string | null;
  transitPort?: string | null;
  temperature?: string | null;
  vent?: Prisma.Decimal | null;
  ventUnit?: string | null;
  cusHold: boolean;
  terHold: boolean;
  terHoldReason?: string | null;
  isSpecialWarning: boolean;
  specialWarning?: string | null;
  conditionCode?: string | null;
  classifyCode?: string | null;
  conditionMachineCode?: string | null;
  isTruckBarge?: string | null;
  truckNo?: string | null;
  romoocNo?: string | null;
  isBundled?: number;
  containerNoMaster?: string | null;
  noteCont?: string | null;
  noteDamage?: string | null;
  noteSpecialHandling?: string | null;
  note?: string | null;
  createdBy: string;
}

export interface UpdateContainerProps {
  idTos?: string | null;
  idCont?: string | null;
  idContTos?: string | null;
  vesselKey?: string;
  vesselImvoy?: string | null;
  vesselExvoy?: string | null;
  eta?: Date | null;
  etb?: Date | null;
  etd?: Date | null;
  bargeInKey?: string | null;
  bargeOutKey?: string | null;
  deliveryOrder?: string | null;
  blNo?: string | null;
  bookingNo?: string | null;
  housebillNo?: string | null;
  containerNo?: string;
  classCode?: string;
  operationCode?: string;
  fe?: string;
  containerStatusCode?: string;
  cargoTypeCode?: string;
  commodity?: string | null;
  localSizeType?: string;
  isoSizeType?: string;
  isLocalForeign?: string;
  jobmodeCodeIn?: string | null;
  methodCodeIn?: string | null;
  dateIn?: Date | null;
  jobmodeCodeOut?: string | null;
  methodCodeOut?: string | null;
  dateOut?: Date | null;
  requireSurveyDate?: Date | null;
  repairMoveDate?: Date | null;
  eirInNo?: string | null;
  eirOutNo?: string | null;
  stuffNo?: string | null;
  unstuffNo?: string | null;
  serviceNo?: string | null;
  draftNo?: string | null;
  invoiceNo?: string | null;
  zoneCode?: string | null;
  yardCode?: string | null;
  lineCode?: string | null;
  block?: string | null;
  bay?: string | null;
  row?: string | null;
  tier?: string | null;
  area?: string | null;
  vgm?: boolean;
  mcWeight?: Prisma.Decimal | null;
  tareWeight?: Prisma.Decimal | null;
  maxGrossWeight?: Prisma.Decimal | null;
  sealNo?: string | null;
  sealNo1?: string | null;
  sealNo2?: string | null;
  pol?: string | null;
  pod?: string | null;
  fpod?: string | null;
  transitCode?: string | null;
  transitPort?: string | null;
  temperature?: string | null;
  vent?: Prisma.Decimal | null;
  ventUnit?: string | null;
  cusHold?: boolean;
  terHold?: boolean;
  terHoldReason?: string | null;
  isSpecialWarning?: boolean;
  specialWarning?: string | null;
  conditionCode?: string | null;
  classifyCode?: string | null;
  conditionMachineCode?: string | null;
  isTruckBarge?: string | null;
  truckNo?: string | null;
  romoocNo?: string | null;
  isBundled?: number;
  containerNoMaster?: string | null;
  noteCont?: string | null;
  noteDamage?: string | null;
  noteSpecialHandling?: string | null;
  note?: string | null;
  updatedBy: string;
}